EXPA-05 UPD: token object moved to services
